move tests, refactor structural, add py.typed, add to_file method, add ruff config, reformat all files
