+++ pyproject.toml
@@ -8,7 +8,7 @@ authors = [{ name = "Enzo Agosta", email = "agosta.enzowork@gmail.com" }]
 description = "Python library for manipulating, creating and editing tmx files"
 readme = { file = "README.md", content-type = "text/markdown" }
 license = { text = "MIT" }
-requires-python = ">=3.11"
+requires-python = ">=3.10"
 classifiers = [
   "Programming Language :: Python :: 3.13",
   "License :: OSI Approved :: MIT License",
@@ -18,7 +18,7 @@ dependencies = ["lxml", "typing_extensions"]
 
 [project.optional-dependencies]
 docs = ["sphinx","pydata-sphinx-theme"]
-dev = ["mypy", "types-lxml"]
+dev = ["mypy", "types-lxml", "ruff", "sphinx","pydata-sphinx-theme"]
 
 [project.urls]
 Homepage = "https://github.com/ChonkyYoshi/python-tmx"
